requirements.txt
alice-blue==1.0.2
altair==4.2.0
api-display-purposes==0.0.3
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astor==0.8.1
async-generator==1.10
asyncio==3.4.3
atomicwrites==1.4.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.10.0
bidict==0.21.4
bleach==4.1.0
blinker==1.4
bs4==0.0.1
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
chardet==3.0.4
charset-normalizer==2.0.9
clarifai==2.6.2
click==7.1.2
colorama==0.4.4
configparser==3.8.1
constantly==15.1.0
cryptography==36.0.1
cycler==0.11.0
daphne==3.0.2
debugpy==1.5.1
decorator==4.0.11
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.1
django-admin==1.0.2
djangorestframework==3.13.1
docutils==0.18.1
EasyProcess==0.3
emoji==1.6.1
entrypoints==0.3
Faker==11.3.0
filelock==3.4.2
filetype==1.0.5
finnhub-python==2.4.8
Flask==2.0.2
fonttools==4.28.5
future==0.18.2
fyers-apiv2==2.0.4
gitdb==4.0.9
GitPython==3.1.24
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
gunicorn==20.1.0
h11==0.12.0
httpie==3.0.2
httplib2==0.20.2
huepy==1.2.1
hyperlink==21.0.0
idna==2.10
imageio==2.1.2
imageio-ffmpeg==0.4.5
importlib-metadata==4.11.3
incremental==21.3.0
iniconfig==1.1.1
instabot==0.117.0
instaclient==2.9.20
instagram-private-api==1.6.0.0
instagram-private-api-extensions==0.3.8
instagrapi==1.16.10
instapy==0.6.15
instapy-cli==0.0.12
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.3.3
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==23.5.0
kiwisolver==1.3.2
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
MeaningCloud-python==2.0.0
mistune==0.8.4
mock==4.0.3
moviepy==0.2.3.2
multidict==6.0.2
multitasking==0.0.10
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.6.7
notebook==6.4.6
numpy==1.21.4
outcome==1.1.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.8.2
platformdirs==2.4.1
pluggy==1.0.0
plyer==2.0.0
prettytable==2.4.0
proglog==0.1.9
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protlib==1.4
protobuf==3.19.1
py==1.11.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/Sushma/Downloads/PyAudio-0.2.11-cp310-cp310-win_amd64.whl
pycparser==2.21
pycryptodomex==3.12.0
pydantic==1.8.2
pydeck==0.7.1
Pygments==2.11.1
Pympler==1.0.1
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyrsistent==0.18.0
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.1
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
readme-renderer==34.0
regex==2021.11.10
requests==2.25.1
requests-toolbelt==0.9.1
responses==0.16.0
rfc3986==2.0.0
rsa==4.8
schedule==1.1.0
scikit-learn==1.0.2
scipy==1.7.3
selenium==4.1.0
semantic-version==2.8.5
Send2Trash==1.8.0
service-identity==21.1.0
setuptools-rust==1.1.2
six==1.16.0
sklearn==0.0
smartapi-python==1.2.8
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
stockcodelib @ file:///C:/Users/Sushma/Documents/SmartApi%20Stock%20Code%20Lib/dist/stockcodelib-0.4-py3-none-any.whl
streamlit==1.3.1
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.63.0
tradingview-ta==3.2.10
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
tweet-preprocessor==0.6.0
twine==3.8.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.7
validators==0.18.2
virtualenv==20.13.0
virtualenvwrapper-win==1.2.7
watchdog==2.1.6
wcwidth==0.2.5
webdriverdownloader==1.1.0.3
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.3
whitenoise==5.3.0
widgetsnbextension==3.5.2
wsproto==1.0.0
yfinance==0.1.68
zestful-parse-ingredient==0.0.7
zipp==3.7.0
zope.interface==5.4.0
